day15: fix algo
So. The manhattan heuristic was wrong, we might need to backtrack.

BUT: we shouldn't backtrack in an unbounded way. As shown in the
backtracking test, going around a 9 might be useful and get you a path
costing 5. But you can't backtrack indefinitely and get a shortest path.

I'm sure there must be an optimal backtracking budget between 7 and 12.
Just to be sure I put it higher at 20, and it works.