day15: correct implementation with breadth-first search instead of depth first
This uses a binary heap to alway go towards the shortest path. First
path to reach the end is considered winner, like in Dijkstra's
algorithm.

Initial inspiration was from A* (remember manhattan), but it was
fundamentally wrong because manhattan assumes cost to go to adjascent
cells is uniform, just that cells might be forbidden. Here we are close
to a Dijkstra implementation.